Unify the way we count expended states between checkers

+++ src/mc/VisitedState.hpp
@@ -20,8 +20,8 @@ public:
   std::shared_ptr<simgrid::mc::Snapshot> system_state = nullptr;
   std::size_t heap_bytes_used = 0;
   int actors_count            = 0;
-  int num          = 0; // unique id of that state in the storage of all stored IDs
-  int original_num = -1; // num field of the VisitedState to which I was declared equal to (used for dot_output)
+  long num                                            = 0; // unique id of that state in the storage of all stored IDs
+  long original_num = -1; // num field of the VisitedState to which I was declared equal to (used for dot_output)
 
   explicit VisitedState(unsigned long state_number);
 };
